- #110: Moving war tier types to folder.
