+++ src/app/hanzhong/war-tiers/index.ts
@@ -1,2 +1,3 @@
 export { HanzhongWarTiersLayout } from './Layout';
 export { HANZHONG_WAR_TIERS } from './war-tiers--data';
+export type { HanzhongWarTierType } from './WarTierType';
